+++ changelog.md
@@ -6,11 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 > Started changelog at the end of the project lifespan, so we begin at [0.8.0].
 ## Unreleased
 ## [0.8.0] - 2022-02-21
-### Fixed
-- Brand link correctly sends user to the top of the page.
 ### Added
 - Added offsets to nav links for better spacing.
 - Added gradient background to sidebar.
+- Added socials tab next to contact section.
+- Created 'blog__display' class for display screenshots that better adjust to mobile view.
+- Added "Matrix Package for Java" project.
 ### Changed
 - Updated brand icon with a new design.
 - Renamed *'filler'* class to *'sidebar'* for better readability.
+- Projects icons are now requested from [DevIcon](https://devicon.dev/) instead of the 'react-icons' library.
+### Fixed
+- Brand link correctly sends user to the top of the page.